# Conflicts:
#   app/src/main/res/values/strings.xml

+++ app/src/main/java/com/globallogic/thespaceapp/data/remote/response/RoadsterResponse.kt
@@ -23,7 +23,7 @@ data class RoadsterResponse(
     @SerializedName("inclination")
     val inclination: Double,
     @SerializedName("launch_date_unix")
-    val launchDateUnix: Int,
+    val launchDateUnix: Long,
     @SerializedName("launch_date_utc")
     val launchDateUtc: String,
     @SerializedName("launch_mass_kg")
